clean up models rerun migrations now working on sa-to-pydantic
