+++ alembic/script.py.mako
@@ -7,6 +7,7 @@ Create Date: ${create_date}
 """
 from alembic import op
 import sqlalchemy as sa
+import sqlalchemy_utils
 ${imports if imports else ""}
 
 # revision identifiers, used by Alembic.
