+++ alembic/versions/1f768d1d7a6a_init.py
@@ -1,0 +1,83 @@
+"""init
+
+Revision ID: 1f768d1d7a6a
+Revises: 
+Create Date: 2021-08-14 12:30:38.040119
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlalchemy_utils
+
+
+# revision identifiers, used by Alembic.
+revision = '1f768d1d7a6a'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('course',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('course_name', sa.String(), nullable=False),
+    sa.Column('is_draft', sa.Boolean(), nullable=True),
+    sa.Column('is_published', sa.Boolean(), nullable=True),
+    sa.Column('date_published', sa.Date(), nullable=True),
+    sa.Column('last_updated', sa.Date(), nullable=True),
+    sa.Column('author', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('course_sections',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('course_section_name', sa.String(), nullable=False),
+    sa.Column('date_published', sa.Date(), nullable=True),
+    sa.Column('last_updated', sa.Date(), nullable=True),
+    sa.Column('course_section_is_draft', sa.Boolean(), nullable=True),
+    sa.Column('course_section_is_complete', sa.Boolean(), nullable=True),
+    sa.Column('course_section_purpose', sa.String(), nullable=False),
+    sa.Column('course_section_order', sa.Integer(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('first_name', sa.String(), nullable=False),
+    sa.Column('last_name', sa.String(), nullable=False),
+    sa.Column('is_admin', sa.Boolean(), nullable=True),
+    sa.Column('user_email', sqlalchemy_utils.types.email.EmailType(length=255), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('authors',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_courses', sa.String(), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('course_sections_association',
+    sa.Column('course_id', sa.Integer(), nullable=False),
+    sa.Column('course_sections_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['course_id'], ['course.id'], ),
+    sa.ForeignKeyConstraint(['course_sections_id'], ['course_sections.id'], ),
+    sa.PrimaryKeyConstraint('course_id', 'course_sections_id')
+    )
+    op.create_table('students',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('enrolled_courses', sa.String(), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('students')
+    op.drop_table('course_sections_association')
+    op.drop_table('authors')
+    op.drop_table('users')
+    op.drop_table('course_sections')
+    op.drop_table('course')
+    # ### end Alembic commands ###
