--- alembic/versions/578ce7104de9_to_camel_case_course_sections.py
@@ -1,95 +1,0 @@
-"""To camel case course_sections
-
-Revision ID: 578ce7104de9
-Revises: df46a9d1d310
-Create Date: 2021-08-11 15:42:15.121046
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "578ce7104de9"
-down_revision = "df46a9d1d310"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "course",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("course_name", sa.String(), nullable=False),
-        sa.Column("is_draft", sa.Boolean(), nullable=True),
-        sa.Column("is_published", sa.Boolean(), nullable=True),
-        sa.Column("date_published", sa.Date(), nullable=True),
-        sa.Column("last_updated", sa.Date(), nullable=True),
-        sa.Column("author", sa.String(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "course_sections",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("course_section_name", sa.String(), nullable=False),
-        sa.Column("date_published", sa.Date(), nullable=True),
-        sa.Column("last_updated", sa.Date(), nullable=True),
-        sa.Column("course_section_is_draft", sa.Boolean(), nullable=True),
-        sa.Column("course_section_is_complete", sa.Boolean(), nullable=True),
-        sa.Column("course_section_purpose", sa.String(), nullable=False),
-        sa.Column("course_section_order", sa.Integer(), nullable=False),
-        sa.Column("course_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["course_id"],
-            ["course.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.drop_table("course-sections")
-    op.drop_table("courses")
-    op.add_column("authors", sa.Column("user_id", sa.Integer(), nullable=True))
-    op.create_foreign_key(None, "authors", "users", ["user_id"], ["id"])
-    op.add_column("students", sa.Column("user_id", sa.Integer(), nullable=True))
-    op.create_foreign_key(None, "students", "users", ["user_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "students", type_="foreignkey")
-    op.drop_column("students", "user_id")
-    op.drop_constraint(None, "authors", type_="foreignkey")
-    op.drop_column("authors", "user_id")
-    op.create_table(
-        "courses",
-        sa.Column("id", sa.INTEGER(), autoincrement=True, nullable=False),
-        sa.Column("course_name", sa.VARCHAR(), autoincrement=False, nullable=False),
-        sa.Column("is_draft", sa.BOOLEAN(), autoincrement=False, nullable=True),
-        sa.Column("is_published", sa.BOOLEAN(), autoincrement=False, nullable=True),
-        sa.Column("date_published", sa.DATE(), autoincrement=False, nullable=True),
-        sa.Column("last_updated", sa.DATE(), autoincrement=False, nullable=True),
-        sa.Column("author", sa.VARCHAR(), autoincrement=False, nullable=False),
-        sa.PrimaryKeyConstraint("id", name="courses_pkey"),
-    )
-    op.create_table(
-        "course-sections",
-        sa.Column(
-            "id",
-            sa.INTEGER(),
-            server_default=sa.text("nextval('\"course-sections_id_seq\"'::regclass)"),
-            autoincrement=True,
-            nullable=False,
-        ),
-        sa.Column("course_section_name", sa.VARCHAR(), autoincrement=False, nullable=False),
-        sa.Column("date_published", sa.DATE(), autoincrement=False, nullable=True),
-        sa.Column("last_updated", sa.DATE(), autoincrement=False, nullable=True),
-        sa.Column("course_section_is_draft", sa.BOOLEAN(), autoincrement=False, nullable=True),
-        sa.Column("course_section_is_complete", sa.BOOLEAN(), autoincrement=False, nullable=True),
-        sa.Column("course_section_purpose", sa.VARCHAR(), autoincrement=False, nullable=False),
-        sa.Column("course_section_order", sa.INTEGER(), autoincrement=False, nullable=False),
-        sa.PrimaryKeyConstraint("id", name="course-sections_pkey"),
-    )
-    op.drop_table("course_sections")
-    op.drop_table("course")
-    # ### end Alembic commands ###
