--- alembic/versions/5bb878cb5c7e_users_and_courses_initial.py
@@ -1,47 +1,0 @@
-"""Users and Courses initial
-
-Revision ID: 5bb878cb5c7e
-Revises: 
-Create Date: 2021-08-08 19:27:27.604936
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils as sau
-
-
-# revision identifiers, used by Alembic.
-revision = '5bb878cb5c7e'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('first_name', sa.String(), nullable=False),
-    sa.Column('last_name', sa.String(), nullable=False),
-    sa.Column('is_admin', sa.Boolean(), nullable=True),
-    sa.Column('user_email', sau.types.email.EmailType(length=255), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('courses',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('course_name', sa.String(), nullable=False),
-    sa.Column('is_draft', sa.Boolean(), nullable=True),
-    sa.Column('is_published', sa.Boolean(), nullable=True),
-    sa.Column('date_published', sa.Date(), nullable=True),
-    sa.Column('last_updated', sa.Date(), nullable=True),
-    sa.Column('author', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('courses')
-    op.drop_table('users')
-    # ### end Alembic commands ###
