--- alembic/versions/7bad15df0f29_rename_table.py
@@ -1,42 +1,0 @@
-"""rename table
-
-Revision ID: 7bad15df0f29
-Revises: c81da288b679
-Create Date: 2021-08-11 18:15:13.396997
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '7bad15df0f29'
-down_revision = 'c81da288b679'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('course_sections_association',
-    sa.Column('course_id', sa.Integer(), nullable=False),
-    sa.Column('course_sections_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['course_id'], ['course.id'], ),
-    sa.ForeignKeyConstraint(['course_sections_id'], ['course_sections.id'], ),
-    sa.PrimaryKeyConstraint('course_id', 'course_sections_id')
-    )
-    op.drop_table('association')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('association',
-    sa.Column('course_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('course_sections_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['course_id'], ['course.id'], name='association_course_id_fkey'),
-    sa.ForeignKeyConstraint(['course_sections_id'], ['course_sections.id'], name='association_course_sections_id_fkey'),
-    sa.PrimaryKeyConstraint('course_id', 'course_sections_id', name='association_pkey')
-    )
-    op.drop_table('course_sections_association')
-    # ### end Alembic commands ###
