--- alembic/versions/c81da288b679_many_to_many.py
@@ -1,38 +1,0 @@
-"""many to many
-
-Revision ID: c81da288b679
-Revises: 578ce7104de9
-Create Date: 2021-08-11 17:31:00.673262
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'c81da288b679'
-down_revision = '578ce7104de9'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('association',
-    sa.Column('course_id', sa.Integer(), nullable=False),
-    sa.Column('course_sections_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['course_id'], ['course.id'], ),
-    sa.ForeignKeyConstraint(['course_sections_id'], ['course_sections.id'], ),
-    sa.PrimaryKeyConstraint('course_id', 'course_sections_id')
-    )
-    op.drop_constraint('course_sections_course_id_fkey', 'course_sections', type_='foreignkey')
-    op.drop_column('course_sections', 'course_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('course_sections', sa.Column('course_id', sa.INTEGER(), autoincrement=False, nullable=True))
-    op.create_foreign_key('course_sections_course_id_fkey', 'course_sections', 'course', ['course_id'], ['id'])
-    op.drop_table('association')
-    # ### end Alembic commands ###
