--- alembic/versions/df46a9d1d310_create_course_sections_and_other_nested_.py
@@ -1,50 +1,0 @@
-"""Create Course Sections and other nested Models. Not altogether working as wanted. Some placeholder text. See comments in files.
-
-Revision ID: df46a9d1d310
-Revises: 5bb878cb5c7e
-Create Date: 2021-08-08 23:27:29.504246
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'df46a9d1d310'
-down_revision = '5bb878cb5c7e'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('authors',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created_courses', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('students',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('enrolled_courses', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('course-sections',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('course_section_name', sa.String(), nullable=False),
-    sa.Column('date_published', sa.Date(), nullable=True),
-    sa.Column('last_updated', sa.Date(), nullable=True),
-    sa.Column('course_section_is_draft', sa.Boolean(), nullable=True),
-    sa.Column('course_section_is_complete', sa.Boolean(), nullable=True),
-    sa.Column('course_section_purpose', sa.String(), nullable=False),
-    sa.Column('course_section_order', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('course-sections')
-    op.drop_table('students')
-    op.drop_table('authors')
-    # ### end Alembic commands ###
