+++ models/users.py
@@ -6,34 +6,6 @@ from sqlalchemy.sql.schema import Column, ForeignKey
 from database.db import Base
 
 
-class Author(Base):
-    __tablename__ = "authors"
-    id = Column(Integer, primary_key=True)
-    created_courses = Column(String, nullable=True)  # This is a placeholder for a list
-    ## Is this correct?
-    user_id = Column(Integer, ForeignKey("users.id"))
-    user = relationship("User", back_populates="authors")
-
-
-class Student(Base):
-    __tablename__ = "students"
-    id = Column(Integer, primary_key=True)
-    enrolled_courses = Column(String, nullable=True)  # This is a placeholder for a list
-    ## Is this correct?
-    user_id = Column(Integer, ForeignKey("users.id"))
-    user = relationship("User", back_populates="students")
-
-
-# What about this?
-# class Admin(Base):
-#     __tablename__ = 'admins'
-#     id = Column(Integer, primary_key=True)
-#     permission_level = Column(Integer, default=1)
-## Is this correct?
-# user_id = Column(Integer, ForeignKey("users.id"))
-# user = relationship("User", back_populates="admins")
-
-
 class User(Base):
     __tablename__ = "users"
 
@@ -43,7 +15,18 @@ class User(Base):
     is_admin = Column(Boolean, default=False)
     user_email = Column(EmailType, nullable=False)
 
-    # Is this correct?
-    authors = relationship("Author", back_populates="user")
-    students = relationship("Student", back_populates="user")
-    # admins = relationship("Admin", back_populates="users")
+
+class Author(Base):
+    __tablename__ = "authors"
+    id = Column(Integer, primary_key=True)
+    created_courses = Column(String, nullable=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    user = relationship("User", back_populates="authors")
+
+
+class Student(Base):
+    __tablename__ = "students"
+    id = Column(Integer, primary_key=True)
+    enrolled_courses = Column(String, nullable=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    user = relationship("User", back_populates="students")
