+++ schemas/users.py
@@ -2,6 +2,10 @@ from typing import Optional
 
 from pydantic import BaseModel
 from pydantic.networks import EmailStr
+from pydantic_sqlalchemy import sqlalchemy_to_pydantic
+
+from models.users import User
+from models.courses import Course
 
 
 class Author(BaseModel):
@@ -12,10 +16,8 @@ class Student(BaseModel):
     enrolled_courses: str # This is a placeholder for a list
 
 
-# class Admin(BaseModel):
-#     permission_level: int
-
-
+CreateUser = sqlalchemy_to_pydantic(User)
+"""
 class CreateUser(BaseModel):
     first_name: str
     last_name: str
@@ -24,3 +26,4 @@ class CreateUser(BaseModel):
     author: Optional[Author] = None
     student: Optional[Student] = None
     # Admin: Optional[Admin] = None
+"""
